+++ tests/test_dates/test_dates_arrays.py
@@ -11,7 +11,16 @@ def pytest_generate_tests(metafunc):
     if "date_func" in metafunc.fixturenames:
         metafunc.parametrize(
             "date_func",
-            ["pandas", "pydatetime", "numpy", "pandas_tz", "pydatetime_tz"],
+            [
+                "pandas",
+                "pydatetime",
+                "numpy",
+                "pandas_tz",
+                "pydatetime_tz",
+                "pandas_timedelta",
+                "pytimedelta",
+                "numpy_timedelta",
+            ],
             indirect=True,
         )
 
@@ -35,6 +44,12 @@ def date_func(request):
                 ts, pytz.timezone("Australia/Sydney")
             ).to_pydatetime()
         )
+    elif request.param == "pandas_timedelta":
+        return lambda ts: ts - pd.Timestamp(2019, 12, 31)
+    elif request.param == "pytimedelta":
+        return lambda ts: (ts - pd.Timestamp(2019, 12, 31)).to_pytimedelta()
+    elif request.param == "numpy_timedelta":
+        return lambda ts: (ts - pd.Timestamp(2019, 12, 31)).to_timedelta64()
     else:
         assert False, "should not happen"